Feature newrelic integration

+++ RITracking/app/src/main/java/de/rocketinternet/android/tracking/trackers/RIAdjustTracker.java
@@ -23,8 +23,8 @@ import de.rocketinternet.android.tracking.utils.RILogUtils;
 
 /**
  * @author alessandro.balocco
- *         <p/>
- *         Convenience controller to proxy-pass tracking information to AdJust
+ *
+ * Convenience controller to proxy-pass tracking information to AdJust
  */
 public class RIAdjustTracker extends RITracker implements
         RIEventTracking,
@@ -59,7 +59,7 @@ public class RIAdjustTracker extends RITracker implements
     }
 
     @Override
-    public void trackEvent(String event, int value, String action, String category, Map<String, Object> data) {
+    public void trackEvent(String event, long value, String action, String category, Map<String, Object> data) {
         RILogUtils.logDebug("Adjust tracker - Tracking event with name: " + event);
 
         if (data == null) {
